DispersiveShallowWater.jl API

DispersiveShallowWater.bathymetry_flat — Constant

bathymetry_flat = DispersiveShallowWater.BathymetryFlat()

A singleton struct indicating a flat bathymetry.

DispersiveShallowWater.boundary_condition_periodic — Constant

boundary_condition_periodic = DispersiveShallowWater.BoundaryConditionPeriodic()

A singleton struct indicating periodic boundary conditions.

DispersiveShallowWater.boundary_condition_reflecting — Constant

boundary_condition_reflecting = DispersiveShallowWater.BoundaryConditionReflecting()

A singleton struct indicating reflecting boundary conditions.

DispersiveShallowWater.AbstractEquations — Type

AbstractEquations{NDIMS, NVARS}

An abstract supertype of specific equations such as the BBM-BBM equations. The type parameters encode the number of spatial dimensions (NDIMS) and the number of primary variables (NVARS) of the physics model.

See also AbstractShallowWaterEquations.

DispersiveShallowWater.AbstractShallowWaterEquations — Type

AbstractShallowWaterEquations{NDIMS, NVARS}

An abstract supertype of all equation system that contain the classical shallow water equations as a subsystem, e.g., the BBMBBMEquations1D, the SvaerdKalischEquations1D, and the SerreGreenNaghdiEquations1D. In 1D, the shallow water equations with flat bathymetry are given by

\[egin{aligned} h_t + (h v)_x &= 0,\ h v_t + rac{1}{2} g (h^2)_x + rac{1}{2} h (v^2)_x &= 0, nd{aligned}\]

where $h$ is the waterheight, $v$ the velocity, and $g$ the gravity_constant.

DispersiveShallowWater.AbstractSolver — Type

AbstractSolver

An abstract supertype of specific solvers.

DispersiveShallowWater.AnalysisCallback — Type

AnalysisCallback(semi; interval=0,
                       extra_analysis_errors=Symbol[],
                       extra_analysis_integrals=(),
                       io=stdout)

Analyze a numerical solution every interval time steps. The L2- and the L∞-norm for each component are computed by default. Additional errors can be computed, e.g. by passing extra_analysis_errors = (:conservation_error,).

Further scalar functions func in extra_analysis_integrals are applied to the numerical solution and integrated over the computational domain. Some examples for this are entropy, and energy_total. You can also write your own function with the same signature as the examples listed above and pass it via extra_analysis_integrals. The computed errors and intergrals are saved for each timestep and can be obtained by calling errors and integrals.

During the Simulation, the AnalysisCallback will print information to io.

DispersiveShallowWater.BBMBBMEquations1D — Type

BBMBBMEquations1D(; gravity_constant, D = 1.0, eta0 = 0.0)

BBM-BBM (Benjamin–Bona–Mahony) system in one spatial dimension. The equations are given by

\[\begin{aligned} \eta_t + ((\eta + D)v)_x - \frac{1}{6}D^2\eta_{xxt} &= 0,\\ v_t + g\eta_x + \left(\frac{1}{2}v^2\right)_x - \frac{1}{6}D^2v_{xxt} &= 0. \end{aligned}\]

The unknown quantities of the BBM-BBM equations are the total water height $\eta$ and the velocity $v$. The gravitational constant is denoted by g and the constant bottom topography (bathymetry) $b = \eta_0 - D$. The water height above the bathymetry is therefore given by $h = \eta - \eta_0 + D$. The BBM-BBM equations are only implemented for $\eta_0 = 0$.

One reference for the BBM-BBM system can be found in Bona et al. (1998). The semidiscretization implemented here conserves the mass and the energy and is developed in Ranocha et al. (2020).

Jerry L. Bona, Min Chen (1998) A Boussinesq system for two-way propagation of nonlinear dispersive waves DOI: 10.1016/S0167-2789(97)00249-2

Hendrik Ranocha, Dimitrios Mitsotakis, David I. Ketcheson (2020) A Broad Class of Conservative Numerical Methods for Dispersive Wave Equations DOI: 10.4208/cicp.OA-2020-0119

DispersiveShallowWater.BBMBBMVariableEquations1D — Type

BBMBBMVariableEquations1D(; gravity_constant, eta0 = 0.0)

BBM-BBM (Benjamin–Bona–Mahony) system in one spatial dimension with spatially varying bathymetry. The equations are given by

\[\begin{aligned} \eta_t + ((\eta + D)v)_x - \frac{1}{6}(D^2\eta_{xt})_x &= 0,\\ v_t + g\eta_x + \left(\frac{1}{2}v^2\right)_x - \frac{1}{6}(D^2v_t)_{xx} &= 0. \end{aligned}\]

The unknown quantities of the BBM-BBM equations are the total water height $\eta$ and the velocity $v$. The gravitational constant is denoted by g and the bottom topography (bathymetry) $b = \eta_0 - D$. The water height above the bathymetry is therefore given by $h = \eta - \eta_0 + D$. The BBM-BBM equations are only implemented for $\eta_0 = 0$.

One reference for the BBM-BBM system with spatially varying bathymetry can be found in Israwi et al. (2022). The semidiscretization implemented here conserves the mass and the energy, is well-balanced for the lake-at-rest state, and is developed in Lampert and Ranocha (2024).

Samer Israwi, Henrik Kalisch, Theodoros Katsaounis, Dimitrios Mitsotakis (2022) A regularized shallow-water waves system with slip-wall boundary conditions in a basin: theory and numerical analysis DOI: 10.1088/1361-6544/ac3c29

Joshua Lampert, Hendrik Ranocha (2024) Structure-Preserving Numerical Methods for Two Nonlinear Systems of Dispersive Wave Equations DOI: 10.48550/arXiv.2402.16669

DispersiveShallowWater.Mesh1D — Type

Mesh1D

Struct that holds the information for a simple homogeneous one-dimensional mesh.

DispersiveShallowWater.Mesh1D — Method

Mesh1D(xmin, xmax, N)

Create a simple homogeneous one-dimensional mesh from xmin to xmax with N nodes.

DispersiveShallowWater.RelaxationCallback — Type

RelaxationCallback(invariant)

Use a relaxation method in time in order to exactly preserve the (nonlinear) invariant for a conservative semidiscretization. A possible choice for invariant is invariant = entropy.

Reference

Hendrik Ranocha, Mohammed Sayyari, Lisandro Dalcin, Matteo Parsani, David I. Ketcheson (2020) Relaxation Runge–Kutta Methods: Fully-Discrete Explicit Entropy-Stable Schemes for the Compressible Euler and Navier–Stokes Equations DOI: 10.1137/19M1263480

DispersiveShallowWater.Semidiscretization — Type

Semidiscretization

A struct containing everything needed to describe a spatial semidiscretization of an equation.

DispersiveShallowWater.Semidiscretization — Method

Semidiscretization(mesh, equations, initial_condition, solver;
                   source_terms=nothing,
                   boundary_conditions=boundary_condition_periodic,
                   RealT=real(solver),
                   uEltype=RealT,
                   initial_cache=(tmp1 = Array{RealT}(undef, nnodes(mesh)),))

Construct a semidiscretization of a PDE.

DispersiveShallowWater.SerreGreenNaghdiEquations1D — Type

SerreGreenNaghdiEquations1D(; gravity_constant, eta0 = 0.0)

Serre-Green-Naghdi system in one spatial dimension. The equations for flat bathymetry are given by

\[\begin{aligned} h_t + (h v)_x &= 0,\\ h v_t - \frac{1}{3} (h^3 v_{tx})_x + \frac{1}{2} g (h^2)_x + \frac{1}{2} h (v^2)_x + p_x &= 0,\\ p &= \frac{1}{3} h^3 v_{x}^2 - \frac{1}{3} h^3 v v_{xx}. \end{aligned}\]

The unknown quantities of the Serre-Green-Naghdi equations are the total water height $\eta = h + b$ and the velocity $v$. The gravitational constant is denoted by g and the bottom topography (bathymetry) $b = \eta_0 - D$. The water height above the bathymetry is therefore given by $h = \eta - \eta_0 + D$. The Serre-Green-Naghdi equations are only implemented for $\eta_0 = 0$. The total water height is therefore given by $\eta = h + b$.

References for the Serre-Green-Naghdi system can be found in

Serre (1953) Contribution â l'étude des écoulements permanents et variables dans les canaux DOI: 10.1051/lhb/1953034

Green and Naghdi (1976) A derivation of equations for wave propagation in water of variable depth DOI: 10.1017/S0022112076002425

The semidiscretization implemented here conserves

the total water mass (integral of h) as a linear invariant

the total momentum (integral of h v) as a nonlinear invariant

the total energy

for periodic boundary conditions, see

Hendrik Ranocha and Mario Ricchiuto (2024) Structure-preserving approximations of the Serre-Green-Naghdi equations in standard and hyperbolic form arXiv: 2408.02665

DispersiveShallowWater.Solver — Type

Solver

A struct that holds the summation by parts (SBP) operators that are used for the spatial discretization.

DispersiveShallowWater.Solver — Method

Solver(mesh, accuracy_order)

Create a solver, where the summation by parts (SBP) operators are of order accuracy_order and associated to the mesh.

DispersiveShallowWater.Solver — Method

Solver(D1, D2)

Create a solver, where D1 is an AbstractDerivativeOperator from SummationByPartsOperators.jl of first derivative_order and D2 is an AbstractDerivativeOperator of second derivative_order or an AbstractMatrix. It can also be nothing if no second derivative is used by the discretization. Both summation-by-parts operators should be associated with the same grid.

DispersiveShallowWater.SummaryCallback — Type

SummaryCallback(io::IO = stdout)

Create and return a callback that resets the timer at the beginning of a simulation and prints the timer values at the end of the simulation.

DispersiveShallowWater.SvaerdKalischEquations1D — Type

SvaerdKalischEquations1D(; gravity_constant, eta0 = 0.0,
                           alpha = 0.0,
                           beta = 0.2308939393939394,
                           gamma = 0.04034343434343434)

Dispersive system by Svärd and Kalisch in one spatial dimension with spatially varying bathymetry. The equations are given in conservative variables by

\[\begin{aligned} h_t + (hv)_x &= (\hat\alpha(\hat\alpha(h + b)_x)_x)_x,\\ (hv)_t + (hv^2)_x + gh(h + b)_x &= (\hat\alpha v(\hat\alpha(h + b)_x)_x)_x + (\hat\beta v_x)_{xt} + \frac{1}{2}(\hat\gamma v_x)_{xx} + \frac{1}{2}(\hat\gamma v_{xx})_x, \end{aligned}\]

where $\hat\alpha^2 = \alpha\sqrt{gD}D^2$, $\hat\beta = \beta D^3$, $\hat\gamma = \gamma\sqrt{gD}D^3$. The coefficients $\alpha$, $\beta$ and $\gamma$ are provided in dimensionless form and $D = \eta_0 - b$ is the still-water depth and eta0 is the still-water surface (lake-at-rest). The equations can be rewritten in primitive variables as

\[\begin{aligned} \eta_t + ((\eta + D)v)_x = (\hat\alpha(\hat\alpha\eta_x)_x)_x,\\ v_t(\eta + D) - v((\eta + D)v)_x + ((\eta + D)v^2)_x + g(\eta + D)\eta_x &= (\hat\alpha v(\hat\alpha\eta_x)_x)_x - v(\hat\alpha(\hat\alpha\eta_x)_x)_x + (\hat\beta v_x)_{xt} + \frac{1}{2}(\hat\gamma v_x)_{xx} + \frac{1}{2}(\hat\gamma v_{xx})_x. \end{aligned}\]

The unknown quantities of the Svärd-Kalisch equations are the total water height $\eta$ and the velocity $v$. The gravitational constant is denoted by g and the bottom topography (bathymetry) $b = \eta_0 - D$. The water height above the bathymetry is therefore given by $h = \eta - \eta_0 + D$.

SvärdKalischEquations1D is an alias for SvaerdKalischEquations1D.

The equations by Svärd and Kalisch are presented and analyzed in Svärd and Kalisch (2023). The semidiscretization implemented here conserves the mass and the energy, is well-balanced for the lake-at-rest state, and is developed in Lampert and Ranocha (2024).

Magnus Svärd, Henrik Kalisch (2023) A novel energy-bounded Boussinesq model and a well-balanced and stable numerical discretization arXiv: 2302.09924

Joshua Lampert, Hendrik Ranocha (2024) Structure-Preserving Numerical Methods for Two Nonlinear Systems of Dispersive Wave Equations DOI: 10.48550/arXiv.2402.16669

DispersiveShallowWater.bathymetry — Function

bathymetry(q, equations)

Return the bathymetry of the primitive variables q for a given set of equations.

q is a vector of the primitive variables at a single node, i.e., a vector of the correct length nvariables(equations).

DispersiveShallowWater.cons2prim — Function

cons2prim(u, equations)

Convert the conserved variables u to the primitive variables for a given set of equations. u is a vector type of the correct length nvariables(equations). Notice the function doesn't include any error checks for the purpose of efficiency, so please make sure your input is correct. The inverse conversion is performed by prim2cons.

DispersiveShallowWater.convergence_test — Method

convergence_test([mod::Module=Main,] example::AbstractString, iterations; io::IO = stdout, kwargs...)
convergence_test([mod::Module=Main,] example::AbstractString, Ns::AbstractVector; io::IO = stdout, kwargs...)

Run multiple simulations using the setup given in example and compute the experimental order of convergence (EOC) in the $L^2$ and $L^\infty$ norm. If iterations is passed as integer, in each iteration, the resolution of the respective mesh will be doubled. If Ns is passed as vector, the simulations will be run for each value of Ns. Additional keyword arguments kwargs... and the optional module mod are passed directly to trixi_include.

Adjusted from Trixi.jl.

DispersiveShallowWater.default_analysis_errors — Method

default_analysis_errors(equations)

Default analysis errors used by the AnalysisCallback.

DispersiveShallowWater.default_analysis_integrals — Method

default_analysis_integrals(equations)

Default analysis integrals used by the AnalysisCallback.

DispersiveShallowWater.default_example — Method

default_example()

Return the path to an example that can be used to quickly see DispersiveShallowWater.jl in action. See also examples_dir and get_examples.

Copied from Trixi.jl.

DispersiveShallowWater.discharge — Method

discharge(q, equations)

See momentum.

DispersiveShallowWater.eachvariable — Method

eachvariable(equations::AbstractEquations)

Return an iterator over the indices that specify the location in relevant data structures for the variables in equations. In particular, not the variables themselves are returned.

DispersiveShallowWater.energy_total — Method

energy_total(q, equations)

Return the total energy of the primitive variables q for a given set of equations. For all AbstractShallowWaterEquations, the total energy is given by the sum of the kinetic and potential energy of the shallow water subsystem, i.e.,

\[\frac{1}{2} h v^2 + \frac{1}{2} g \eta^2\]

in 1D, where $h$ is the waterheight, $\eta = h + b$ the waterheight_total, $v$ the velocity, and $g$ the gravity_constant.

q is a vector of the primitive variables at a single node, i.e., a vector of the correct length nvariables(equations).

DispersiveShallowWater.energy_total_modified — Method

energy_total_modified(q_global, equations::AbstractShallowWaterEquations, cache)

Return the modified total energy of the primitive variables q_global for the equations. This modified total energy is a conserved quantity and can contain additional terms compared to the usual energy_total. For example, for the SvaerdKalischEquations1D and the SerreGreenNaghdiEquations1D, it contains additional terms depending on the derivative of the velocity $v_x$ modeling non-hydrostatic contributions.

q_global is a vector of the primitive variables at ALL nodes. cache needs to hold the SBP operators used by the solver if non-hydrostatic terms are present.

DispersiveShallowWater.energy_total_modified — Method

energy_total_modified(q_global, equations::SerreGreenNaghdiEquations1D, cache)

Return the modified total energy of the primitive variables q_global for the SerreGreenNaghdiEquations1D. It contains an additional term containing a derivative compared to the usual energy_total modeling non-hydrostatic contributions. The energy_total_modified is a conserved quantity (for periodic boundary conditions).

For a bathymetry_flat the total energy is given by

\[\frac{1}{2} g h^2 + \frac{1}{2} h v^2 + \frac{1}{6} h^3 v_x^2.\]

q_global is a vector of the primitive variables at ALL nodes. cache needs to hold the SBP operators used by the solver.

DispersiveShallowWater.energy_total_modified — Method

energy_total_modified(q_global, equations::SvaerdKalischEquations1D, cache)

Return the modified total energy of the primitive variables q_global for the SvaerdKalischEquations1D. It contains an additional term containing a derivative compared to the usual energy_total. The energy_total_modified is a conserved quantity of the Svärd-Kalisch equations.

q_global is a vector of the primitive variables at ALL nodes. cache needs to hold the first-derivative SBP operator D1.

DispersiveShallowWater.entropy — Function

entropy(q, equations)

Return the entropy of the primitive variables q for a given set of equations. For all AbstractShallowWaterEquations, the entropy is just the energy_total.

q is a vector of the primitive variables at a single node, i.e., a vector of the correct length nvariables(equations).

DispersiveShallowWater.entropy_modified — Method

entropy_modified(q_global, equations::AbstractShallowWaterEquations, cache)

Alias for energy_total_modified.

DispersiveShallowWater.errors — Method

errors(analysis_callback)

Return the computed errors for each timestep as a named tuple. The shape of each entry is (nvariables, ntimesteps).

DispersiveShallowWater.examples_dir — Method

examples_dir()

Return the directory where the example files provided with DispersiveShallowWater.jl are located. If DispersiveShallowWater is installed as a regular package (with ]add DispersiveShallowWater), these files are read-only and should not be modified. To find out which files are available, use, e.g., readdir.

Copied from Trixi.jl.

Examples

readdir(examples_dir())

DispersiveShallowWater.get_examples — Method

get_examples()

Return a list of all examples that are provided by DispersiveShallowWater.jl. See also examples_dir and default_example.

Copied from Trixi.jl.

DispersiveShallowWater.get_name — Method

get_name(equations::AbstractEquations)

Return the canonical, human-readable name for the given system of equations.

Examples

julia> DispersiveShallowWater.get_name(BBMBBMEquations1D(gravity_constant=1.0))
"BBMBBMEquations1D"

DispersiveShallowWater.gravity_constant — Method

gravity_constant(equations::AbstractShallowWaterEquations)

Return the gravity constant $g$ for a given set of equations. See also AbstractShallowWaterEquations.

DispersiveShallowWater.initial_condition_convergence_test — Method

initial_condition_convergence_test(x, t, equations::BBMBBMEquations1D, mesh)

A travelling-wave solution used for convergence tests in a periodic domain.

For details see Example 5 in Section 3 from (here adapted for dimensional equations):

Min Chen (1997) Exact Traveling-Wave Solutions to Bidirectional Wave Equations DOI: 10.1023/A:1026667903256

DispersiveShallowWater.initial_condition_convergence_test — Method

initial_condition_convergence_test(x, t, equations::BBMBBMVariableEquations1D, mesh)

A travelling-wave solution used for convergence tests in a periodic domain. The bathymetry is constant.

For details see Example 5 in Section 3 from (here adapted for dimensional equations):

Min Chen (1997) Exact Traveling-Wave Solutions to Bidirectional Wave Equations DOI: 10.1023/A:1026667903256

DispersiveShallowWater.initial_condition_convergence_test — Method

initial_condition_convergence_test(x, t, equations::SerreGreenNaghdiEquations1D, mesh)

A soliton solution used for convergence tests in a periodic domain.

DispersiveShallowWater.initial_condition_dingemans — Method

initial_condition_dingemans(x, t, equations::BBMBBMVariableEquations1D, mesh)

The initial condition that uses the dispersion relation of the Euler equations to approximate waves generated by a wave maker as it is done by experiments of Dingemans. The topography is a trapezoidal.

Translation of water height

The initial condition for the water height is translated to be around 0, which is needed for the simulation because the BBMBBMVariableEquations1D are only implemented for $ta_0 = 0$.

References:

Magnus Svärd, Henrik Kalisch (2023) A novel energy-bounded Boussinesq model and a well-balanced and stable numerical discretization arXiv: 2302.09924

Maarten W. Dingemans (1994) Comparison of computations with Boussinesq-like models and laboratory measurements link

DispersiveShallowWater.initial_condition_dingemans — Method

initial_condition_dingemans(x, t, equations::SvaerdKalischEquations1D, mesh)

The initial condition that uses the dispersion relation of the Euler equations to approximate waves generated by a wave maker as it is done by experiments of Dingemans. The topography is a trapezoidal. It is assumed that equations.eta0 = 0.8.

References:

Magnus Svärd, Henrik Kalisch (2023) A novel energy-bounded Boussinesq model and a well-balanced and stable numerical discretization arXiv: 2302.09924

Maarten W. Dingemans (1994) Comparison of computations with Boussinesq-like models and laboratory measurements link

DispersiveShallowWater.initial_condition_manufactured — Method

initial_condition_manufactured(x, t, equations::BBMBBMEquations1D, mesh)

A smooth manufactured solution in combination with source_terms_manufactured.

DispersiveShallowWater.initial_condition_manufactured — Method

initial_condition_manufactured(x, t, equations::BBMBBMVariableEquations1D, mesh)

A smooth manufactured solution in combination with source_terms_manufactured.

DispersiveShallowWater.initial_condition_manufactured — Method

initial_condition_manufactured(x, t, equations::SvaerdKalischEquations1D, mesh)

A smooth manufactured solution in combination with source_terms_manufactured.

DispersiveShallowWater.initial_condition_manufactured_reflecting — Method

initial_condition_manufactured_reflecting(x, t, equations::BBMBBMEquations1D, mesh)

A smooth manufactured solution for reflecting boundary conditions in combination with source_terms_manufactured_reflecting.

DispersiveShallowWater.initial_condition_manufactured_reflecting — Method

initial_condition_manufactured_reflecting(x, t, equations::BBMBBMVariableEquations1D, mesh)

A smooth manufactured solution for reflecting boundary conditions in combination with source_terms_manufactured_reflecting.

DispersiveShallowWater.integrals — Method

integrals(analysis_callback)

Return the computed integrals for each timestep as a named tuple.

DispersiveShallowWater.momentum — Method

momentum(q, equations)

Return the momentum/discharge of the primitive variables q for a given set of equations, i.e., the waterheight times the velocity.

q is a vector of the primitive variables at a single node, i.e., a vector of the correct length nvariables(equations).

DispersiveShallowWater.prim2cons — Function

prim2cons(q, equations)

Convert the primitive variables q to the conserved variables for a given set of equations. q is a vector type of the correct length nvariables(equations). Notice the function doesn't include any error checks for the purpose of efficiency, so please make sure your input is correct. The inverse conversion is performed by cons2prim.

DispersiveShallowWater.prim2prim — Method

prim2prim(q, equations)

Return the primitive variables q. While this function is as trivial as identity, it is also as useful.

DispersiveShallowWater.semidiscretize — Method

semidiscretize(semi::Semidiscretization, tspan)

Wrap the semidiscretization semi as an ODE problem in the time interval tspan that can be passed to solve from the SciML ecosystem.

DispersiveShallowWater.source_terms_manufactured — Method

source_terms_manufactured(q, x, t, equations::BBMBBMEquations1D, mesh)

A smooth manufactured solution in combination with initial_condition_manufactured.

DispersiveShallowWater.source_terms_manufactured — Method

source_terms_manufactured(q, x, t, equations::BBMBBMVariableEquations1D, mesh)

A smooth manufactured solution in combination with initial_condition_manufactured.

DispersiveShallowWater.source_terms_manufactured — Method

source_terms_manufactured(q, x, t, equations::SvaerdKalischEquations1D, mesh)

A smooth manufactured solution in combination with initial_condition_manufactured.

DispersiveShallowWater.source_terms_manufactured_reflecting — Method

source_terms_manufactured_reflecting(q, x, t, equations::BBMBBMEquations1D, mesh)

A smooth manufactured solution for reflecting boundary conditions in combination with initial_condition_manufactured_reflecting.

DispersiveShallowWater.source_terms_manufactured_reflecting — Method

source_terms_manufactured_reflecting(q, x, t, equations::BBMBBMVariableEquations1D, mesh)

A smooth manufactured solution for reflecting boundary conditions in combination with initial_condition_manufactured_reflecting.

DispersiveShallowWater.still_water_surface — Method

still_water_surface(q, equations::AbstractShallowWaterEquations)

Return the still water surface $\eta_0$ (lake at rest) for a given set of equations.

DispersiveShallowWater.tstops — Method

tstops(analysis_callback)

Return the time values that correspond to the saved values of the errors and integrals.

DispersiveShallowWater.varnames — Function

varnames(conversion_function, equations)

Return the list of variable names when applying conversion_function to the conserved variables associated to equations. Common choices of the conversion_function are prim2prim and prim2cons.

DispersiveShallowWater.velocity — Function

velocity(q, equations)

Return the velocity of the primitive variables q for a given set of equations.

q is a vector of the primitive variables at a single node, i.e., a vector of the correct length nvariables(equations).

DispersiveShallowWater.waterheight — Function

waterheight(q, equations)

Return the waterheight of the primitive variables q for a given set of equations, i.e., the waterheight above the bathymetry.

q is a vector of the primitive variables at a single node, i.e., a vector of the correct length nvariables(equations).

DispersiveShallowWater.waterheight_total — Function

waterheight_total(q, equations)

Return the total waterheight of the primitive variables q for a given set of equations, i.e., the waterheight plus the bathymetry.

q is a vector of the primitive variables at a single node, i.e., a vector of the correct length nvariables(equations).

SummationByPartsOperators.grid — Method

grid(semi)

Get the grid of a semidiscretization.

DispersiveShallowWater.@autoinfiltrate — Macro

@autoinfiltrate
@autoinfiltrate condition::Bool

Invoke the @infiltrate macro of the package Infiltrator.jl to create a breakpoint for ad-hoc interactive debugging in the REPL. If the optional argument condition is given, the breakpoint is only enabled if condition evaluates to true.

As opposed to using Infiltrator.@infiltrate directly, this macro does not require Infiltrator.jl to be added as a dependency to DispersiveShallowWater.jl. As a bonus, the macro will also attempt to load the Infiltrator module if it has not yet been loaded manually.

Note: For this macro to work, the Infiltrator.jl package needs to be installed in your current Julia environment stack.

See also: Infiltrator.jl

Internal use only

Please note that this macro is intended for internal use only. It is not part of the public API of DispersiveShallowWater.jl, and it thus can altered (or be removed) at any time without it being considered a breaking change.