DispersiveShallowWater.jl API

DispersiveShallowWater.boundary_condition_periodic — Constant

boundary_condition_periodic = DispersiveShallowWater.BoundaryConditionPeriodic()

A singleton struct indicating periodic boundary conditions.

DispersiveShallowWater.AbstractEquations — Type

AbstractEquations{NDIMS, NVARS}

An abstract supertype of specific equations such as the BBM-BBM equations. The type parameters encode the number of spatial dimensions (NDIMS) and the number of primary variables (NVARS) of the physics model.

DispersiveShallowWater.AbstractSolver — Type

AbstractSolver

An abstract supertype of specific solvers.

DispersiveShallowWater.AnalysisCallback — Type

AnalysisCallback(semi; interval=0,
                       extra_analysis_errors=Symbol[],
                       extra_analysis_integrals=(),
                       io=stdout)

Analyze a numerical solution every interval time steps. The L2- and the L∞-norm for each component are computed by default. Additional errors can be computed, e.g. by passing extra_analysis_errors = (:conservation_error,).

Further scalar functions func in extra_analysis_integrals are applied to the numerical solution and integrated over the computational domain. Some examples for this are entropy, and energy_total. You can also write your own function with the same signature as the examples listed above and pass it via extra_analysis_integrals. The computed errors and intergrals are saved for each timestep and can be obtained by calling errors and integrals.

During the Simulation, the AnalysisCallback will print information to io.

DispersiveShallowWater.BBMBBMEquations1D — Type

BBMBBMEquations1D(gravity, D)

BBM-BBM (Benjamin–Bona–Mahony) system in one spatial dimension. The equations are given by

\[\begin{aligned} \eta_t + ((\eta + D)v)_x - \frac{1}{6}D^2\eta_{xxt} &= 0,\\ v_t + g\eta_x + \left(\frac{1}{2}v^2\right)_x - \frac{1}{6}D^2v_{xxt} &= 0. \end{aligned}\]

The unknown quantities of the BBM-BBM equations are the total water height $\eta$ and the velocity $v$. The gravitational constant is denoted by g and the constant bottom topography (bathymetry) $b = -D$. The water height above the bathymetry is therefore given by $h = \eta + D$.

One reference for the BBM-BBM system can be found in

Jerry L. Bona, Min Chen (1998) A Boussinesq system for two-way propagation of nonlinear dispersive waves DOI: 10.1016/S0167-2789(97)00249-2

DispersiveShallowWater.BBMBBMVariableEquations1D — Type

BBMBBMVariableEquations1D(gravity, eta0 = 1.0)

BBM-BBM (Benjamin–Bona–Mahony) system in one spatial dimension with spatially varying bathymetry. The equations are given by

\[\begin{aligned} \eta_t + ((\eta + D)v)_x - \frac{1}{6}(D^2\eta_{xt})_x &= 0,\\ v_t + g\eta_x + \left(\frac{1}{2}v^2\right)_x - \frac{1}{6}(D^2v_t)_{xx} &= 0. \end{aligned}\]

The unknown quantities of the BBM-BBM equations are the total water height $\eta$ and the velocity $v$. The gravitational constant is denoted by g and the bottom topography (bathymetry) $b = -D$. The water height above the bathymetry is therefore given by $h = \eta + D$.

One reference for the BBM-BBM system with spatially varying bathymetry can be found in

Samer Israwi, Henrik Kalisch, Theodoros Katsaounis, Dimitrios Mitsotakis (2022) A regularized shallow-water waves system with slip-wall boundary conditions in a basin: theory and numerical analysis DOI: 10.1088/1361-6544/ac3c29

DispersiveShallowWater.Mesh1D — Type

Mesh1D

Struct that holds the information for a simple homogeneous one-dimensional mesh.

DispersiveShallowWater.Mesh1D — Method

Mesh1D(xmin, xmax, N)

Create a simple homogeneous one-dimensional mesh from xmin to xmax with N nodes.

DispersiveShallowWater.RelaxationCallback — Type

RelaxationCallback(invariant)

Use a relaxation method in time in order to exactly preserve the (nonlinear) invariant for a conservative semidiscretization. A possible choice for invariant is invariant = entropy.

Reference

Hendrik Ranocha, Mohammed Sayyari, Lisandro Dalcin, Matteo Parsani, David I. Ketcheson (2020) Relaxation Runge–Kutta Methods: Fully-Discrete Explicit Entropy-Stable Schemes for the Compressible Euler and Navier–Stokes Equations DOI: 10.1137/19M1263480

DispersiveShallowWater.Semidiscretization — Type

Semidiscretization

A struct containing everything needed to describe a spatial semidiscretization of an equation.

DispersiveShallowWater.Semidiscretization — Method

Semidiscretization(mesh, equations, initial_condition, solver;
                   boundary_conditions=boundary_condition_periodic,
                   RealT=real(solver),
                   uEltype=RealT,
                   initial_cache=NamedTuple())

Construct a semidiscretization of a PDE.

DispersiveShallowWater.Solver — Type

Solver

A struct that holds the summation by parts (SBP) operators that are used for the spatial discretization.

DispersiveShallowWater.Solver — Method

Solver(mesh, accuracy_order)

Create a solver, where the summation by parts (SBP) operators are of order accuracy_order and associated to the mesh.

DispersiveShallowWater.Solver — Method

Solver(D1, D2)

Create a solver, where D1 is an AbstractDerivativeOperator from SummationByPartsOperators.jl of first derivative_order and D2 is an AbstractDerivativeOperator of second derivative_order or an AbstractMatrix. Both summation by parts operators should be associated with the same grid.

DispersiveShallowWater.SvaerdKalischEquations1D — Type

SvärdKalischEquations1D(gravity, eta0 = 1.0, alpha = 0.0, beta = 0.2308939393939394, gamma = 0.04034343434343434)

Dispersive system by Svärd and Kalisch in one spatial dimension with spatially varying bathymetry. The equations are given in conservative variables by

\[\begin{aligned} h_t + (hv)_x &= (\hat\alpha(\hat\alpha(h + b)_x)_x)_x,\\ (hv)_t + (hv^2)_x + gh(h + b)_x &= (\hat\alpha v(\hat\alpha(h + b)_x)_x)_x + (\hat\beta v_x)_{xt} + \frac{1}{2}(\hat\gamma v_x)_{xx} + \frac{1}{2}(\hat\gamma v_{xx})_x, \end{aligned}\]

where $\hat\alpha^2 = \alpha\sqrt{gd}d^2$, $\hat\beta = \beta d^3$, $\hat\gamma = \gamma\sqrt{gd}d^3$. The coefficients $\alpha$, $\beta$ and $\gamma$ are provided in dimensionless form and $d = \eta_0 - b$ is the still-water depth and eta0 is the still-water surface (lake-at-rest). The equations can be rewritten in primitive variables as

\[\begin{aligned} \eta_t + ((\eta + D)v)_x = (\hat\alpha(\hat\alpha\eta_x)_x)_x,\\ v_t(\eta + D) - v((\eta + D)v)_x + ((\eta + D)v^2)_x + g(\eta + D)\eta_x &= (\hat\alpha v(\hat\alpha\eta_x)_x)_x - v(\hat\alpha(\hat\alpha\eta_x)_x)_x + (\hat\beta v_x)_{xt} + \frac{1}{2}(\hat\gamma v_x)_{xx} + \frac{1}{2}(\hat\gamma v_{xx})_x. \end{aligned}\]

The unknown quantities of the Svärd-Kalisch equations are the total water height $\eta$ and the velocity $v$. The gravitational constant is denoted by g and the bottom topography (bathymetry) $b = -D$. The water height above the bathymetry is therefore given by $h = \eta + D$.

The equations by Svärd and Kalisch are presented and analyzed in

Magnus Svärd, Henrik Kalisch (2023) A novel energy-bounded Boussinesq model and a well-balanced and stable numerical discretization arXiv: 2302.09924

DispersiveShallowWater.cons2prim — Function

cons2prim(u, equations)

Convert the conserved variables u to the primitive variables for a given set of equations. u is a vector type of the correct length nvariables(equations). Notice the function doesn't include any error checks for the purpose of efficiency, so please make sure your input is correct. The inverse conversion is performed by prim2cons.

DispersiveShallowWater.convergence_test — Method

convergence_test([mod::Module=Main,] example::AbstractString, iterations; kwargs...)

Run iterations simulations using the setup given in example and compute the experimental order of convergence (EOC) in the $L^2$ and $L^\infty$ norm. In each iteration, the resolution of the respective mesh will be doubled. Additional keyword arguments kwargs... and the optional module mod are passed directly to trixi_include.

Adjusted from Trixi.jl.

DispersiveShallowWater.default_analysis_errors — Method

default_analysis_errors(equations)

Default analysis errors used by the AnalysisCallback.

DispersiveShallowWater.default_analysis_integrals — Method

default_analysis_integrals(equations)

Default analysis integrals used by the AnalysisCallback.

DispersiveShallowWater.default_example — Method

default_example()

Return the path to an example that can be used to quickly see DispersiveShallowWater.jl in action. See also examples_dir and get_examples.

Copied from Trixi.jl.

DispersiveShallowWater.discharge — Method

discharge(q, equations)

See momentum.

DispersiveShallowWater.eachvariable — Method

eachvariable(equations::AbstractEquations)

Return an iterator over the indices that specify the location in relevant data structures for the variables in equations. In particular, not the variables themselves are returned.

DispersiveShallowWater.energy_total — Function

energy_total(q, equations)

Return the total energy of the primitive variables q for a given set of equations.

q is a vector of the primitive variables at a single node, i.e., a vector of the correct length nvariables(equations).

DispersiveShallowWater.energy_total_modified — Method

energy_total_modified(q, equations::SvaerdKalischEquations1D, cache)

Return the modified total energy of the primitive variables q for the SvaerdKalischEquations1D. It contains an additional term containing a derivative compared to the usual energy_total. The energy_total_modified is a conserved quantity of the Svärd-Kalisch equations.

q is a vector of the primitive variables at ALL nodes, i.e., a matrix of the correct length nvariables(equations) as first dimension and the number of nodes as length of the second dimension. cache needs to hold the first-derivative SBP operator D1.

DispersiveShallowWater.entropy — Function

entropy(q, equations)

Return the entropy of the primitive variables q for a given set of equations.

q is a vector of the primitive variables at a single node, i.e., a vector of the correct length nvariables(equations).

DispersiveShallowWater.entropy_modified — Method

entropy_modified(q, equations::SvaerdKalischEquations1D, cache)

Alias for energy_total_modified.

DispersiveShallowWater.errors — Method

errors(analysis_callback)

Return the computed errors for each timestep as a named tuple. The shape of each entry is (nvariables, ntimesteps).

DispersiveShallowWater.examples_dir — Method

examples_dir()

Return the directory where the example files provided with DispersiveShallowWater.jl are located. If DispersiveShallowWater is installed as a regular package (with ]add DispersiveShallowWater), these files are read-only and should not be modified. To find out which files are available, use, e.g., readdir.

Copied from Trixi.jl.

Examples

readdir(examples_dir())

DispersiveShallowWater.get_examples — Method

get_examples()

Return a list of all examples that are provided by DispersiveShallowWater.jl. See also examples_dir and default_example.

Copied from Trixi.jl.

DispersiveShallowWater.get_name — Method

get_name(equations::AbstractEquations)

Return the canonical, human-readable name for the given system of equations.

Examples

julia> DispersiveShallowWater.get_name(BBMBBMEquations1D(gravity_constant=1.0))
"BBMBBMEquations1D"

DispersiveShallowWater.initial_condition_convergence_test — Method

initial_condition_convergence_test(x, t, equations::BBMBBMEquations1D, mesh)

A travelling-wave solution used for convergence tests in a periodic domain.

For details see Example 5 in Section 3 from (here adapted for dimensional equations):

Min Chen (1997) Exact Traveling-Wave Solutions to Bidirectional Wave Equations DOI: 10.1023/A:1026667903256

DispersiveShallowWater.initial_condition_convergence_test — Method

initial_condition_convergence_test(x, t, equations::BBMBBMVariableEquations1D, mesh)

A travelling-wave solution used for convergence tests in a periodic domain. The bathymetry is constant.

For details see Example 5 in Section 3 from (here adapted for dimensional equations):

Min Chen (1997) Exact Traveling-Wave Solutions to Bidirectional Wave Equations DOI: 10.1023/A:1026667903256

DispersiveShallowWater.initial_condition_dingemans — Method

initial_condition_dingemans(x, t, equations::BBMBBMVariableEquations1D, mesh)

The initial condition that uses the dispersion relation of the Euler equations to approximate waves generated by a wave maker as it is done by experiments of Dingemans. The topography is a trapesoidal.

References:

Magnus Svärd, Henrik Kalisch (2023) A novel energy-bounded Boussinesq model and a well-balanced and stable numerical discretization arXiv: 2302.09924

Maarten W. Dingemans (1994) Comparison of computations with Boussinesq-like models and laboratory measurements link

DispersiveShallowWater.initial_condition_dingemans — Method

initial_condition_dingemans(x, t, equations::SvaerdKalischEquations1D, mesh)

The initial condition that uses the dispersion relation of the Euler equations to approximate waves generated by a wave maker as it is done by experiments of Dingemans. The topography is a trapesoidal. It is assumed that equations.eta0 = 0.8.

References:

Magnus Svärd, Henrik Kalisch (2023) A novel energy-bounded Boussinesq model and a well-balanced and stable numerical discretization arXiv: 2302.09924

Maarten W. Dingemans (1994) Comparison of computations with Boussinesq-like models and laboratory measurements link

DispersiveShallowWater.integrals — Method

integrals(analysis_callback)

Return the computed integrals for each timestep as a named tuple.

DispersiveShallowWater.momentum — Method

momentum(q, equations)

Return the momentum of the primitive variables q for a given set of equations, i.e. the waterheight times the velocity.

q is a vector of the primitive variables at a single node, i.e., a vector of the correct length nvariables(equations).

DispersiveShallowWater.path_create_figures — Method

path_create_figures()

Return the path to the file that creates all figures used in the master thesis "Structure-preserving Numerical Methods for Dispersive Shallow Water Model" (2023). Executing this julia script may take a while.

Examples

include(DispersiveShallowWater.path_create_figures())

DispersiveShallowWater.prim2cons — Function

prim2cons(q, equations)

Convert the primitive variables q to the conserved variables for a given set of equations. q is a vector type of the correct length nvariables(equations). Notice the function doesn't include any error checks for the purpose of efficiency, so please make sure your input is correct. The inverse conversion is performed by cons2prim.

DispersiveShallowWater.prim2prim — Method

prim2prim(q, equations)

Return the primitive variables q. While this function is as trivial as identity, it is also as useful.

DispersiveShallowWater.semidiscretize — Method

semidiscretize(semi::Semidiscretization, tspan)

Wrap the semidiscretization semi as an ODE problem in the time interval tspan that can be passed to solve from the SciML ecosystem.

DispersiveShallowWater.trixi_include — Method

trixi_include([mod::Module=Main,] example::AbstractString; kwargs...)

include the file example and evaluate its content in the global scope of module mod. You can override specific assignments in example by supplying keyword arguments. It's basic purpose is to make it easier to modify some parameters while running DispersiveShallowWater from the REPL. Additionally, this is used in tests to reduce the computational burden for CI while still providing examples with sensible default values for users.

Before replacing assignments in example, the keyword argument maxiters is inserted into calls to solve and DispersiveShallowWater.solve with it's default value used in the SciML ecosystem for ODEs, see the "Miscellaneous" section of the documentation.

Copied from Trixi.jl.

Examples

julia> redirect_stdout(devnull) do
           trixi_include(@__MODULE__, joinpath(examples_dir(), "bbm_bbm_1d", "bbm_bbm_1d_basic.jl"),
                         tspan=(0.0, 0.1))
           sol.t[end]
       end
0.1

DispersiveShallowWater.tstops — Method

tstops(analysis_callback)

Return the time values that correspond to the saved values of the errors and integrals.

DispersiveShallowWater.varnames — Function

varnames(conversion_function, equations)

Return the list of variable names when applying conversion_function to the conserved variables associated to equations. Common choices of the conversion_function are prim2prim and prim2cons.

DispersiveShallowWater.velocity — Function

velocity(q, equations)

Return the velocity of the primitive variables q for a given set of equations.

q is a vector of the primitive variables at a single node, i.e., a vector of the correct length nvariables(equations).

DispersiveShallowWater.waterheight — Function

waterheight(q, equations)

Return the waterheight of the primitive variables q for a given set of equations, i.e. the waterheight above the bathymetry.

q is a vector of the primitive variables at a single node, i.e., a vector of the correct length nvariables(equations).

DispersiveShallowWater.waterheight_total — Function

waterheight_total(q, equations)

Return the total waterheight of the primitive variables q for a given set of equations, i.e. the waterheight plus the bathymetry.

q is a vector of the primitive variables at a single node, i.e., a vector of the correct length nvariables(equations).

SummationByPartsOperators.grid — Method

grid(semi)

Get the grid of a semidiscretization.

DispersiveShallowWater.@autoinfiltrate — Macro

@autoinfiltrate
@autoinfiltrate condition::Bool

Invoke the @infiltrate macro of the package Infiltrator.jl to create a breakpoint for ad-hoc interactive debugging in the REPL. If the optional argument condition is given, the breakpoint is only enabled if condition evaluates to true.

As opposed to using Infiltrator.@infiltrate directly, this macro does not require Infiltrator.jl to be added as a dependency to DispersiveShallowWater.jl. As a bonus, the macro will also attempt to load the Infiltrator module if it has not yet been loaded manually.

Note: For this macro to work, the Infiltrator.jl package needs to be installed in your current Julia environment stack.

See also: Infiltrator.jl

Internal use only

Please note that this macro is intended for internal use only. It is not part of the public API of DispersiveShallowWater.jl, and it thus can altered (or be removed) at any time without it being considered a breaking change.